# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sourcetool's API documentation",
        "title": "Sourcetool API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "https://api.trysourcetool.com",
    "basePath": "/api/v1.",
    "paths": {
        "/apiKeys": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apiKeys"
                ],
                "operationId": "list-apikeys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListAPIKeysPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apiKeys"
                ],
                "operationId": "create-apikey",
                "parameters": [
                    {
                        "description": " ",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateAPIKeyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateAPIKeyPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            }
        },
        "/apiKeys/{apiKeyID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apiKeys"
                ],
                "operationId": "get-apikey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetAPIKeyPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apiKeys"
                ],
                "operationId": "update-apikey",
                "parameters": [
                    {
                        "description": " ",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateAPIKeyInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateAPIKeyPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apiKeys"
                ],
                "operationId": "delete-apikey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key ID",
                        "name": "apiKeyID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteAPIKeyPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            }
        },
        "/environments": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "operationId": "list-environments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListEnvironmentsPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "operationId": "create-environment",
                "parameters": [
                    {
                        "description": " ",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateEnvironmentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateEnvironmentPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            }
        },
        "/environments/{environmentID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "operationId": "get-environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID",
                        "name": "environmentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetEnvironmentPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "operationId": "update-environment",
                "parameters": [
                    {
                        "description": " ",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateEnvironmentInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Environment ID",
                        "name": "environmentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateEnvironmentPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "operationId": "delete-environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID",
                        "name": "environmentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteEnvironmentPayload"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/errdefs.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errdefs.Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "types.APIKeyPayload": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "environment": {
                    "$ref": "#/definitions/types.EnvironmentPayload"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "types.CreateAPIKeyInput": {
            "type": "object",
            "required": [
                "environmentId",
                "name"
            ],
            "properties": {
                "environmentId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.CreateAPIKeyPayload": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/types.APIKeyPayload"
                }
            }
        },
        "types.CreateEnvironmentInput": {
            "type": "object",
            "required": [
                "color",
                "name",
                "slug"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "types.CreateEnvironmentPayload": {
            "type": "object",
            "properties": {
                "environment": {
                    "$ref": "#/definitions/types.EnvironmentPayload"
                }
            }
        },
...
        "types.DeleteAPIKeyPayload": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/types.APIKeyPayload"
                }
            }
        },
        "types.DeleteEnvironmentPayload": {
            "type": "object",
            "properties": {
                "environment": {
                    "$ref": "#/definitions/types.EnvironmentPayload"
                }
            }
        },
...
        "types.EnvironmentPayload": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "types.GetAPIKeyPayload": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/types.APIKeyPayload"
                }
            }
        },
        "types.GetEnvironmentPayload": {
            "type": "object",
            "properties": {
                "environment": {
                    "$ref": "#/definitions/types.EnvironmentPayload"
                }
            }
        },
...
        "types.ListAPIKeysPayload": {
            "type": "object",
            "properties": {
                "devKey": {
                    "$ref": "#/definitions/types.APIKeyPayload"
                },
                "liveKeys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.APIKeyPayload"
                    }
                }
            }
        },
        "types.ListEnvironmentsPayload": {
            "type": "object",
            "properties": {
                "environments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.EnvironmentPayload"
                    }
                }
            }
        },
...
        "types.UpdateAPIKeyInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "types.UpdateAPIKeyPayload": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "$ref": "#/definitions/types.APIKeyPayload"
                }
            }
        },
        "types.UpdateEnvironmentInput": {
            "type": "object",
            "required": [
                "color",
                "name"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.UpdateEnvironmentPayload": {
            "type": "object",
            "properties": {
                "environment": {
                    "$ref": "#/definitions/types.EnvironmentPayload"
                }
            }
        },
...
    }
}